+++ src/main.rs
@@ -3,7 +3,7 @@ use std::net::TcpListener;
 use clap::Parser;
 use servers::{
     plugins::loader,
-    tcp::{handle_connection, Client},
+    tcp::{handle_connection, handle_websocket, Client},
 };
 use simple_logger::SimpleLogger;
 
@@ -29,9 +29,19 @@ struct Cli {
         display_order = 2
     )]
     port: String,
+
+    #[clap(
+        short = 'w',
+        long = "ws-port",
+        default_value = "9998",
+        help = "Websocket server port [set 0 to random]",
+        display_order = 2
+    )]
+    ws_port: String,
 }
 
-fn main() -> anyhow::Result<()> {
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
     // init logger
     SimpleLogger::new().init()?;
 
@@ -39,14 +49,18 @@ fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
 
     // start tcp server
-    start_server(&cli.host, &cli.port)?;
+    tokio::spawn(start_ws_server(
+        cli.host.clone(),
+        cli.ws_port,
+        cli.port.clone(),
+    ));
+    start_tcp_server(cli.host, cli.port).await?;
 
     Ok(())
 }
 
 /// Start tcp server
-#[tokio::main]
-async fn start_server(host: &str, port: &str) -> anyhow::Result<()> {
+async fn start_tcp_server(host: String, port: String) -> anyhow::Result<()> {
     // listen Tcp server
     let listener = TcpListener::bind(format!("{host}:{port}"))?;
 
@@ -61,6 +75,23 @@ async fn start_server(host: &str, port: &str) -> anyhow::Result<()> {
 
         // handle client connection in new thread
         tokio::spawn(handle_connection(client, plugin_manager.clone()));
+    }
+
+    // server for a unexpectedly reason be terminated
+    panic!("Server unexpectedly terminated!")
+}
+
+/// Start WebSocket server
+async fn start_ws_server(host: String, port: String, tcp_port: String) -> anyhow::Result<()> {
+    // listen Tcp server
+    let listener = tokio::net::TcpListener::bind(format!("{host}:{port}")).await?;
+
+    println!("WebSocket server started at: {}", listener.local_addr()?);
+
+    // Accepts a new incoming connection from this listener.
+    while let Ok((stream, _address)) = listener.accept().await {
+        let tcp_port = tcp_port.clone();
+        tokio::spawn(handle_websocket(stream, tcp_port));
     }
 
     // server for a unexpectedly reason be terminated