+++ src/tcp/mod.rs
@@ -2,6 +2,8 @@
 
 mod client;
 mod handle_connection;
+mod handle_websocket;
 
 pub use client::*;
 pub use handle_connection::*;
+pub use handle_websocket::*;